Move js file to resource directory; migrate Plone 4.
