+++ collective/js/throttledebounce/configure.zcml
@@ -3,10 +3,13 @@
     xmlns:browser="http://namespaces.zope.org/browser"
     xmlns:five="http://namespaces.zope.org/five"
     xmlns:genericsetup="http://namespaces.zope.org/genericsetup"
+    xmlns:upgrade-step="http://namespaces.zope.org/ftw.upgrade"
     xmlns:zcml="http://namespaces.zope.org/zcml"
     i18n_domain="collective.js.throttledebounce">
 
     <five:registerPackage package="." />
+
+    <include package="ftw.upgrade" file="meta.zcml" />
 
     <configure zcml:condition="not-have plone-5">
       <genericsetup:registerProfile
@@ -46,8 +49,14 @@
           />
     </configure>
 
-    <browser:resource
-        name="jquery.ba-throttle-debounce.js"
-        file="jquery.ba-throttle-debounce.js" />
+    <upgrade-step:directory
+        profile="collective.js.throttledebounce:default"
+        directory="upgrades"
+        />
+
+    <browser:resourceDirectory
+        name="collective.js.throttledebounce"
+        directory="resources"
+        />
 
 </configure>
