+++ collective/js/throttledebounce/profiles/default/metadata.xml
@@ -1,4 +1,3 @@
 <?xml version="1.0"?>
 <metadata>
-  <version>1.0</version>
 </metadata>
