+++ collective/js/throttledebounce/upgrades/20191217181055_move_js_to_resource_folder_for_plone_4/jsregistry.xml
@@ -1,6 +1,8 @@
 <?xml version="1.0"?>
 <object name="portal_javascripts" meta_type="JavaScripts Registry">
 
+  <javascript id="++resource++jquery.ba-throttle-debounce.js" remove="True" />
+
   <javascript
       id="++resource++collective.js.throttledebounce/throttledebounce.js"
       cacheable="True"
