+++ collective/js/throttledebounce/upgrades/20191217181055_move_js_to_resource_folder_for_plone_4/upgrade.py
@@ -1,0 +1,14 @@
+from ftw.upgrade import UpgradeStep
+from Products.CMFPlone.utils import getFSVersionTuple
+
+
+IS_PLONE_4 = getFSVersionTuple() < (5, )
+
+
+class MoveJsToResourceFolderForPlone4(UpgradeStep):
+    """Move js to resource folder for Plone 4.
+    """
+
+    def __call__(self):
+        if IS_PLONE_4:
+            self.install_upgrade_profile()
